poll for debug events, don't register a handler
Registering an event handler from a running application makes that event
handler go stale and presumably crash when a service calls it after the
app doesn't exist anymore.

Switch to polling for debug events. This also simplifies logic where the
service had to keep track of the handler and whether to generate the
events.